extracted updatedSince logic
