+++ packages/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './window';
 export * from './numbers';
+export * from './date';
